+++ src/main/scala/millfork/compiler/CompilationContext.scala
@@ -7,6 +7,12 @@ import millfork.env.{Environment, MangledFunction, NormalFunction}
   * @author Karol Stasiak
   */
 case class CompilationContext(env: Environment, function: NormalFunction, extraStackOffset: Int, options: CompilationOptions){
+  def withInlinedEnv(environment: Environment): CompilationContext = {
+    val newEnv = new Environment(Some(env), MfCompiler.nextLabel("en"))
+    newEnv.things ++= environment.things
+    copy(env = newEnv)
+  }
+
 
   def addStack(i: Int): CompilationContext = this.copy(extraStackOffset = extraStackOffset + i)