+++ CHANGELOG.md
@@ -2,9 +2,19 @@
 
 ## Current version
 
+* **Breaking change!** Renamed `inline` to `macro`
+
+* Added support for parameters for macros written in Millfork
+
+* Added optimizer hints: `inline`, `noinline`, `register`
+
+* Added `*'=` and `<<<<` operators
+
 * Added return dispatch statements.
 
 * Fixed several optimization bugs.
+
+* Fixed several C64 library bugs.
 
 * Other minor improvements.